+++ tensorflow_datasets/core/features/audio_feature.py
@@ -17,10 +17,15 @@
 
 from __future__ import annotations
 
+import abc
+import enum
+import functools
+import io
 import os
-from typing import Optional, Union
+from typing import BinaryIO, Optional, Union
 import wave
 
+from absl import logging
 from etils import epath
 import numpy as np
 from tensorflow_datasets.core import lazy_imports_lib
@@ -33,6 +38,157 @@ from tensorflow_datasets.core.utils.lazy_imports_utils import tensorflow as tf
 
 Json = type_utils.Json
 Encoding = tensor_feature.Encoding
+
+
+class _TfioFileFormat(enum.Enum):
+  """Format of the audio files supported for decoding by `tensorflow_io`.
+
+  Attributes:
+    FLAC: Free Lossless Audio Codec
+  """
+  FLAC = 'flac'
+
+
+@functools.lru_cache(maxsize=None)
+def _tfio_decode_fn():
+  tfio = lazy_imports_lib.lazy_imports.tensorflow_io
+
+  return {
+      _TfioFileFormat.FLAC: tfio.audio.decode_flac,
+  }
+
+
+@functools.lru_cache(maxsize=None)
+def _tfio_acceptable_dtypes():
+  return {
+      _TfioFileFormat.FLAC: [tf.uint8, tf.int16, tf.int32],
+  }
+
+
+class _AudioDecoder(abc.ABC):
+  """Utils which encode/decode audios."""
+
+  def __init__(self, file_format: Optional[str], np_dtype: np.dtype,
+               shape: utils.Shape):
+    """Constructs the lazy audio decoder.
+
+    Args:
+      file_format: `str`, the audio file format. Can be any format ffmpeg
+        understands. If in `_TfioFileFormat`, then will use `tensorflow_io`
+      np_dtype: The numpy dtype of the data.
+      shape: `tuple`, shape of the data.
+    """
+    self._file_format = file_format
+    self._np_dtype = np_dtype
+    self._dtype = tf.dtypes.as_dtype(self._np_dtype)
+    self._shape = shape
+
+  @abc.abstractmethod
+  def encode_audio(self, fobj: BinaryIO,
+                   file_format: Optional[str]) -> np.ndarray:
+    """Encode audio into numpy array for storing as a tf-example."""
+    raise NotImplementedError
+
+  @abc.abstractmethod
+  def decode_audio(self, audio_tensor: tf.Tensor) -> tf.Tensor:
+    """Decode audio from the loaded tf-example."""
+    raise NotImplementedError
+
+
+def _pydub_load_audio(fobj: BinaryIO, file_format: Optional[str]) -> np.ndarray:
+  """Read audio using pydub library."""
+  pydub = lazy_imports_lib.lazy_imports.pydub
+
+  audio_segment = pydub.AudioSegment.from_file(fobj, format=file_format)
+  raw_samples = np.array(audio_segment.get_array_of_samples())
+
+  if audio_segment.channels > 1:
+    return raw_samples.reshape((-1, audio_segment.channels))
+  else:
+    return raw_samples
+
+
+@tf.function(
+    input_signature=(
+        tf.TensorSpec(shape=(), dtype=tf.string),
+        tf.OptionalSpec(tf.TensorSpec(shape=(), dtype=tf.string)),
+    ))
+def _tf_decode_audio(audio_tensor: tf.Tensor,
+                     file_format_tensor: tf.Tensor) -> tf.Tensor:
+  """Decode audio from tf.Tensor."""
+
+  def pydub_decode(audio_tensor: tf.Tensor,
+                   file_format_tensor: tf.Tensor) -> np.ndarray:
+    fobj = io.BytesIO(audio_tensor.numpy())
+    if file_format_tensor.has_value():
+      file_format = file_format_tensor.get_value().numpy()
+    else:
+      file_format = None
+    return _pydub_load_audio(fobj, file_format)
+
+  # pydub.AudioSegment.get_array_of_samples returns an array with type code `b`,
+  # `h` or `i` which can be all converted to `tf.int32`
+  result = tf.py_function(pydub_decode, [audio_tensor, file_format_tensor],
+                          tf.int32)
+
+  return result
+
+
+class _LazyDecoder(_AudioDecoder):
+  """Read audio during decoding."""
+
+  def __init__(self, *args, **kwargs):
+    super().__init__(*args, **kwargs)
+
+    try:
+      self._tfio_file_format = _TfioFileFormat(self._file_format)
+    except ValueError:
+      self._tfio_file_format = None
+      logging.warning(
+          'Using lazy encoding with `file_format=%s` might be very slow when '
+          'reading prepared dataset. Consider using one of these file formats: '
+          '%s', self._file_format, [item.value for item in _TfioFileFormat])
+
+    if self._tfio_file_format:
+      acceptable_dtypes = _tfio_acceptable_dtypes()[self._tfio_file_format]
+      if self._dtype not in acceptable_dtypes:
+        raise ValueError(
+            f'Acceptable `dtype` for lazy loading {self._file_format}: '
+            f'{acceptable_dtypes} (was {self._dtype})')
+
+  def encode_audio(self, fobj: BinaryIO,
+                   file_format: Optional[str]) -> np.ndarray:
+    return np.array(fobj.read(), dtype=tf.string.as_numpy_dtype)
+
+  def decode_audio(self, audio_tensor: tf.Tensor) -> tf.Tensor:
+    if self._tfio_file_format:
+      audio_tensor = _tfio_decode_fn()[self._tfio_file_format](
+          audio_tensor, dtype=self._dtype)
+      audio_tensor = tf.squeeze(audio_tensor)
+    else:
+      if self._file_format:
+        file_format_tensor = tf.experimental.Optional.from_value(
+            self._file_format)
+      else:
+        file_format_tensor = tf.experimental.Optional.empty(
+            tf.TensorSpec(shape=(), dtype=tf.string))
+      audio_tensor = _tf_decode_audio(audio_tensor, file_format_tensor)
+
+    audio_tensor.set_shape(self._shape)
+
+    return tf.cast(audio_tensor, self._dtype)
+
+
+class _EagerDecoder(_AudioDecoder):
+  """Read audio during encoding."""
+
+  def encode_audio(self, fobj: BinaryIO,
+                   file_format: Optional[str]) -> np.ndarray:
+    audio = _pydub_load_audio(fobj, file_format)
+    return audio.astype(self._np_dtype)
+
+  def decode_audio(self, audio_tensor: tf.Tensor) -> tf.Tensor:
+    return audio_tensor
 
 
 class Audio(tensor_feature.Tensor):
@@ -57,6 +213,7 @@ class Audio(tensor_feature.Tensor):
       sample_rate: Optional[int] = None,
       encoding: Union[str, Encoding] = Encoding.NONE,
       doc: feature_lib.DocArg = None,
+      lazy_decode: bool = False,
   ):
     """Constructs the connector.
 
@@ -71,29 +228,38 @@ class Audio(tensor_feature.Tensor):
       encoding: Internal encoding. See `tfds.features.Encoding` for available
         values.
       doc: Documentation of this feature (e.g. description).
+      lazy_decode: `bool`, if set `True` then stores audio as is and decodes it
+        to numpy array when loaded. Otherwise saves decoded audio.
     """
     dtype = tf.dtypes.as_dtype(dtype)
     self._file_format = file_format
+    self._sample_rate = sample_rate
     if len(shape) > 2:
       raise ValueError('Audio shape should be either (length,) or '
                        f'(length, num_channels), got {shape}.')
-    self._shape = shape
-    self._sample_rate = sample_rate
-    super().__init__(shape=shape, dtype=dtype, encoding=encoding, doc=doc)
 
-  def _encode_file(self, fobj, file_format):
-    audio_segment = lazy_imports_lib.lazy_imports.pydub.AudioSegment.from_file(
-        fobj, format=file_format)
-    np_dtype = np.dtype(self.np_dtype)
-    raw_samples = np.array(audio_segment.get_array_of_samples())
-    raw_samples = raw_samples.astype(np_dtype)
-    if audio_segment.channels > 1:
-      audio_data = raw_samples.reshape((-1, audio_segment.channels))
+    if lazy_decode:
+      serialized_dtype = tf.string
+      serialized_shape = ()
+      _Decoder = _LazyDecoder
     else:
-      audio_data = raw_samples
-    return super().encode_example(audio_data)
+      serialized_dtype = None
+      serialized_shape = None
+      _Decoder = _EagerDecoder
+
+    super().__init__(
+        shape=shape,
+        dtype=dtype,
+        encoding=encoding,
+        doc=doc,
+        serialized_dtype=serialized_dtype,
+        serialized_shape=serialized_shape)
+
+    self._audio_decoder = _Decoder(
+        file_format=self._file_format, np_dtype=self._dtype, shape=self._shape)
 
   def encode_example(self, audio_or_path_or_fobj):
+    """Convert the given audio into a dict convertible to tf example."""
     if isinstance(audio_or_path_or_fobj, (np.ndarray, list)):
       return audio_or_path_or_fobj
     elif isinstance(audio_or_path_or_fobj, epath.PathLikeCls):
@@ -101,11 +267,19 @@ class Audio(tensor_feature.Tensor):
       file_format = self._file_format or filename.split('.')[-1]
       with tf.io.gfile.GFile(filename, 'rb') as audio_f:
         try:
-          return self._encode_file(audio_f, file_format)
+          audio = self._audio_decoder.encode_audio(audio_f, file_format)
         except Exception as e:  # pylint: disable=broad-except
           utils.reraise(e, prefix=f'Error for {filename}: ')
     else:
-      return self._encode_file(audio_or_path_or_fobj, self._file_format)
+      audio = self._audio_decoder.encode_audio(audio_or_path_or_fobj,
+                                               self._file_format)
+
+    return super().encode_example(audio)
+
+  def decode_example(self, tfexample_data):
+    """See base class for details."""
+    audio = super().decode_example(tfexample_data)
+    return self._audio_decoder.decode_audio(audio)
 
   @property
   def sample_rate(self):