+++ tensorflow_datasets/core/features/audio_feature_test.py
@@ -30,16 +30,9 @@ from tensorflow_datasets.core import features
 class AudioFeatureTest(testing.FeatureExpectationsTestCase,
                        parameterized.TestCase):
 
-  @parameterized.parameters([
-      (np.int64, 1),
-      (tf.int64, 1),
-      (np.int64, 2),
-      (tf.int64, 2),
-      (np.int64, 8),
-      (tf.int64, 8),
-  ])
+  @parameterized.product(dtype=[np.int64, tf.int64], num_channels=[1, 2, 8])
   def test_numpy_array(self, dtype, num_channels):
-    np_audio = _create_np_audio(num_channels)
+    np_audio = _create_np_audio(num_channels, np.int64)
 
     self.assertFeature(
         feature=features.Audio(
@@ -57,10 +50,10 @@ class AudioFeatureTest(testing.FeatureExpectationsTestCase,
             sample_rate=1000,
         ))
 
-  @parameterized.parameters([(np.float32, 1), (tf.float32, 1), (np.float32, 2),
-                             (tf.float32, 2), (np.float32, 8), (tf.float32, 8)])
+  @parameterized.product(dtype=[np.float32, tf.float32], num_channels=[1, 2, 8])
   def test_numpy_array_float(self, dtype, num_channels):
-    np_audio = _create_np_audio(num_channels).astype(np.float32)
+    np_audio = _create_np_audio(num_channels, np.float32)
+
     self.assertFeature(
         feature=features.Audio(
             dtype=dtype, shape=_shape_for_channels(num_channels)),
@@ -74,17 +67,26 @@ class AudioFeatureTest(testing.FeatureExpectationsTestCase,
         ],
     )
 
-  @parameterized.parameters([(1,), (2,), (8,)])
-  def test_wav_file(self, num_channels):
-    np_audio = _create_np_audio(num_channels)
+  @parameterized.product(
+      dtype=[np.int16, np.int32],
+      lazy_decode=[True, False],
+      num_channels=[1, 2, 8],
+  )
+  def test_wav_file(self, dtype, lazy_decode, num_channels):
+    file_format = 'wav'
+
+    np_audio = _create_np_audio(num_channels, dtype)
     _, tmp_file = tempfile.mkstemp()
-    _write_wave_file(np_audio, tmp_file)
+    _write_audio_file(np_audio, tmp_file, file_format)
 
     self.assertFeature(
         feature=features.Audio(
-            file_format='wav', shape=_shape_for_channels(num_channels)),
+            file_format=file_format,
+            shape=_shape_for_channels(num_channels),
+            dtype=dtype,
+            lazy_decode=lazy_decode),
         shape=_shape_for_channels(num_channels),
-        dtype=np.int64,
+        dtype=dtype,
         tests=[
             testing.FeatureExpectationItem(
                 value=tmp_file,
@@ -95,13 +97,51 @@ class AudioFeatureTest(testing.FeatureExpectationsTestCase,
                 expected=np_audio,
             ),
         ],
-        test_attributes=dict(_file_format='wav',))
+        test_attributes=dict(_file_format=file_format,))
 
-  @parameterized.parameters([(1,), (2,), (8,)])
-  def test_file_object(self, num_channels):
-    np_audio = _create_np_audio(num_channels)
+  @parameterized.product(lazy_decode=[True, False], num_channels=[1, 2, 8])
+  def test_flac_file(self, lazy_decode, num_channels):
+    if lazy_decode:
+      if 'dev' in tf.__version__:
+        self.skipTest('`tensorflow_io` is not compatible with `tf-nightly`')
+
+    dtype = np.int16
+    file_format = 'flac'
+
+    np_audio = _create_np_audio(num_channels, dtype)
     _, tmp_file = tempfile.mkstemp()
-    _write_wave_file(np_audio, tmp_file)
+    _write_audio_file(np_audio, tmp_file, file_format)
+
+    self.assertFeature(
+        feature=features.Audio(
+            file_format=file_format,
+            shape=_shape_for_channels(num_channels),
+            dtype=dtype,
+            lazy_decode=lazy_decode),
+        shape=_shape_for_channels(num_channels),
+        dtype=dtype,
+        tests=[
+            testing.FeatureExpectationItem(
+                value=tmp_file,
+                expected=np_audio,
+            ),
+            testing.FeatureExpectationItem(
+                value=pathlib.Path(tmp_file),
+                expected=np_audio,
+            ),
+        ],
+        test_attributes=dict(_file_format=file_format,))
+
+  @parameterized.product(
+      dtype=[np.int16, np.int32],
+      lazy_decode=[True, False],
+      num_channels=[1, 2, 8])
+  def test_file_object(self, dtype, lazy_decode, num_channels):
+    file_format = 'wav'
+
+    np_audio = _create_np_audio(num_channels, dtype)
+    _, tmp_file = tempfile.mkstemp()
+    _write_audio_file(np_audio, tmp_file, file_format)
 
     class GFileWithSeekOnRead(tf.io.gfile.GFile):
       """Wrapper around GFile which is reusable across multiple read() calls.
@@ -118,9 +158,12 @@ class AudioFeatureTest(testing.FeatureExpectationsTestCase,
     with GFileWithSeekOnRead(tmp_file, 'rb') as file_obj:
       self.assertFeature(
           feature=features.Audio(
-              file_format='wav', shape=_shape_for_channels(num_channels)),
+              file_format=file_format,
+              shape=_shape_for_channels(num_channels),
+              dtype=dtype,
+              lazy_decode=lazy_decode),
           shape=_shape_for_channels(num_channels),
-          dtype=tf.int64,
+          dtype=dtype,
           tests=[
               testing.FeatureExpectationItem(
                   value=file_obj,
@@ -132,31 +175,51 @@ class AudioFeatureTest(testing.FeatureExpectationsTestCase,
   def test_sample_rate_property(self):
     self.assertEqual(features.Audio(sample_rate=1600).sample_rate, 1600)
 
+  def test_lazy_decode_error_file(self):
+    with self.assertRaisesWithPredicateMatch(
+        ValueError, 'Acceptable `dtype` for lazy loading'):
+      features.Audio(file_format='flac', lazy_decode=True)
+
 
 def _shape_for_channels(num_channels, *, length=None):
   """Returns the shape."""
   return (length,) if num_channels == 1 else (length, num_channels)
 
 
-def _create_np_audio(num_channels: int) -> np.ndarray:
+def _create_np_audio(num_channels: int, dtype: np.dtype) -> np.ndarray:
   """Creates a random audio `np.array`."""
-  return np.random.randint(
-      -2**10,
-      2**10,
-      size=_shape_for_channels(num_channels, length=10),
-      dtype=np.int64)
+  if np.issubdtype(dtype, np.integer):
+    dtype_info = np.iinfo(dtype)
+  else:
+    dtype_info = np.finfo(dtype)
+
+  np_audio = np.random.randint(
+      max(-2**10, int(dtype_info.min)),
+      min(2**10, int(dtype_info.max)),
+      size=_shape_for_channels(num_channels, length=10))
+
+  return np_audio.astype(dtype)
 
 
-def _write_wave_file(np_audio, path):
-  """Creates a random audio file."""
+def _write_audio_file(np_audio, path, file_format):
+  """Creates a random audio file with the given file format."""
+  dtype = np_audio.dtype
+
+  if (file_format == 'flac' and
+      dtype not in [np.int16]) or (file_format == 'wav' and
+                                   dtype not in [np.int16, np.int32]):
+    raise ValueError(
+        f'Can\'t save audio in "{file_format.upper()}" format with dtype '
+        f'`{np_audio.dtype}`.')
+
+  data = array.array(np.sctype2char(dtype), np_audio.reshape((-1,)))
+  sample_width = np.dtype(dtype).alignment
   num_channels = np_audio.shape[1] if len(np_audio.shape) == 2 else 1
-  audio = pydub.AudioSegment(
-      b'',
-      sample_width=2,
+
+  pydub.AudioSegment(
+      data=data,
+      sample_width=sample_width,
       channels=num_channels,
       frame_rate=1,
-  )
-  # See documentation for _spawn usage:
-  # https://github.com/jiaaro/pydub/blob/master/API.markdown#audiosegmentget_array_of_samples
-  audio = audio._spawn(array.array(audio.array_type, np_audio.reshape((-1,))))
-  audio.export(path, format='wav')
+  ).export(
+      path, format=file_format)